Improve monorepo support
